Gets layout out of LaunchScreenActivity and into LibraryListFragment
Problem is, I still can't figure out how to get references to the imageViews in the fragment.
It's because I need the activity reference to use getViewByID, but I can't use the constructor
call for that, because storing it in a variable leaks memory, and calling the getViewById methods
in the constructor means that the layout hasn't been instantiated yet, so those methods return null.